changed dmg from int to float

+++ BusinessLayer/Trade.cs
@@ -11,7 +11,7 @@ namespace MonsterTradingCardGame.BusinessLayer
         public int TradeId { get; set; }
         public Card OfferedCard { get; set; }
         public string RequestedCardType { get; set; }
-        public int RequestedMinimumDamage { get; set; }
+        public float RequestedMinimumDamage { get; set; }
         public string Status { get; set; }
     }
 }
